--- example/middleware/Aureolin.ts
@@ -1,9 +1,0 @@
-import { Context, NextFunction } from '../../src'
-
-export const Aureolin = () => {
-    return async (ctx: Context, next: NextFunction) => {
-        ctx.res.setHeader('X-Aureolin', 'Aureolin')
-        ctx.res.setHeader('X-Aureolin-Version', (await import('../../package.json')).version)
-        await next()
-    }
-}
