+++ examples/basic/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "aureolin-basic-example",
+  "version": "1.0.0",
+  "description": "Aureolin Example",
+  "main": "dist/main.js",
+  "scripts": {
+    "dev": "NODE_ENV=development ts-node src/main.ts",
+    "start": "NODE_ENV=production node dist/main.js",
+    "build": "tsc -p .",
+    "prebuild": "rimraf dist"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/alensaito1/aureolin.git"
+  },
+  "keywords": [
+    "example",
+    "aureolin"
+  ],
+  "author": "Alen Yohannnan",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/alensaito1/aureolin/issues"
+  },
+  "homepage": "https://github.com/alensaito1/aureolin#readme",
+  "devDependencies": {
+    "@types/node": "^16.11.6",
+    "@types/react": "^17.0.33",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.4.4"
+  },
+  "dependencies": {
+    "aureolin": "github:alensaito1/aureolin#config-init"
+  }
+}
