+++ examples/basic/src/main.ts
@@ -1,0 +1,8 @@
+import { create } from 'aureolin'
+
+const main = async () => {
+    const app = await create()
+    app.on('error', console.error)
+}
+
+main()
