+++ examples/basic/tsconfig.json
@@ -1,0 +1,24 @@
+{
+    "compilerOptions": {
+        "target": "es6",
+        "lib": [
+            "ESNext"
+        ],                         
+        "module": "commonjs",
+        "outDir": "./dist",                        
+        "resolveJsonModule": true,
+        "strict": true,                           
+        "esModuleInterop": true,                  
+        "skipLibCheck": true,
+        "jsx": "react",                     
+        "forceConsistentCasingInFileNames": true,
+        "experimentalDecorators": true  
+    },
+    "include": [
+        "src/**/*"
+    ],
+    "exclude": [
+        "node_modules",
+        "tests/**/*"
+    ]
+}
